Modified Changed stuff around for the build system
Changed things such that the build system can straight up import the
src directory as a subdirectory. Changed the example CMake to do this
moved the fonts into their own directory and included that in the build
system. Added the start of a widget system to the ssd1306 display
(developed for other projects).